requirements.txt
transformers
torch
faiss-cpu
flask
pillow
requests
numpy
pandas
ollama
markdown
pytest